Move runtime genesis config to presets
