+++ Cargo.toml
@@ -270,6 +270,7 @@ polkadot-node-core-provisioner = { version = "18.0.0" }
 polkadot-node-core-pvf-checker = { version = "18.0.0" }
 polkadot-node-core-runtime-api = { version = "18.0.0" }
 polkadot-statement-distribution = { version = "18.0.0" }
+polkadot-service = { version = "19.0.1" }
 
 polkadot-core-primitives = { version = "15.0.0" }
 polkadot-node-core-parachains-inherent = { version = "18.0.0" }
