+++ relay-node/test/service/src/chain_spec.rs
@@ -20,11 +20,11 @@ use babe_primitives::AuthorityId as BabeId;
 use grandpa::AuthorityId as GrandpaId;
 use pallet_staking::Forcing;
 use polkadot_primitives::{AccountId, AssignmentId, ValidatorId, MAX_CODE_SIZE, MAX_POV_SIZE};
+use polkadot_service::chain_spec::{get_account_id_from_seed, get_from_seed};
 use sc_chain_spec::{ChainSpec, ChainSpecExtension, ChainType};
 pub use sc_consensus_grandpa as grandpa;
 use sc_sync_state_rpc::LightSyncStateExtension;
 use serde::{Deserialize, Serialize};
-use service::chain_spec::{get_account_id_from_seed, get_from_seed};
 use sp_authority_discovery::AuthorityId as AuthorityDiscoveryId;
 pub use sp_consensus_babe as babe_primitives;
 use sp_core::sr25519;
